Update merge_rate_data.py
Joint by taxon and taxon order. Use outer join instead of inner to select the union of taxa and order instead of intersections.

The reason of doing this is that when I used inner join for the whole RTL9 protein, the number of shared taxa equals zero. i.e., none of them have a amino acid substitution rate at every amino acid position. So we have to use outer join.